requirements.txt
accelerate
calflops==0.3.2
diffusers==0.33.1
einops==0.8.1
huggingface_hub==0.30.2
matplotlib==3.10.3
numpy==1.23.0
omegaconf==2.3.0
pandas==2.2.3
Pillow==11.2.1
piq==0.8.0
pyiqa==0.1.13
scikit_learn==1.6.1
scipy==1.7.3
setuptools==75.8.0
tensorboard==2.14.0
thop
torch==2.6.0
torch_fidelity==0.3.0
torchinfo==1.8.0
torchmetrics==1.7.1
torchvision==0.21.0
tqdm==4.67.1
webdataset==0.2.100
